+++ packages/dropdown-menu/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pikas-ui/dropdown-menu
 
+## 1.0.28
+
+### Patch Changes
+
+- Fix portalled props
+
 ## 1.0.27
 
 ### Patch Changes